Component likes created
