+++ src/components/Post.js
@@ -1,19 +1,18 @@
 import styled from "styled-components";
 import { colors } from "../globalStyles";
-import { FiHeart } from "react-icons/fi";
 import { Link } from "react-router-dom";
 import ReactHashtag from "react-hashtag";
 import { MdModeEdit, MdDelete } from "react-icons/md";
 import UserContext from "../contexts/UserContext";
 import { useContext, useState, useRef, useEffect } from "react";
-import { putEdit, postLike } from "../service/api.service";
-import { FaHeart } from "react-icons/fa";
-import { postUnlike } from "../service/api.service";
-import ReactTooltip from "react-tooltip";
+import { putEdit } from "../service/api.service";
+
+
 
 import DeleteModal from "./DeleteModal";
 import CommentIcon from "./CommentIcon";
 import Comments from "./Comments";
+import Likes from "./Likes";
 
 export default function Post(props) {
   const { id, text, link, linkTitle, linkDescription, linkImage, user, likes } =
@@ -24,15 +23,6 @@ export default function Post(props) {
   const [newText, setNewText] = useState(text);
   const [editDisabled, setEditDisabled] = useState(false);
   const inputRef = useRef();
-  const [usersLikesArray, setUsersLikesArray] = useState([
-    ...likes.map((user) => user.userId),
-  ]);
-  const [toolTipMsg, setToolTipMsg] = useState("");
-  const [like, setLike] = useState(
-    usersLikesArray.includes(userData.user.id) ? true : false
-  );
-  const [likesArrayLength, setLikesArrayLength] = useState(likes.length);
-  const [actualLikes, setActualLikes] = useState(likes);
   const [modalOpen, setModalOpen] = useState(false);
   const [showComments, setShowComments] = useState(false);
   const [postComments, setPostComments] = useState([]);
@@ -69,91 +59,6 @@ export default function Post(props) {
         });
     }
   }
-
-  useEffect(() => {
-    setUsersLikesArray([...likes.map((user) => user.userId)]);
-  }, [likes]);
-
-  function isliked() {
-    if (!like) {
-      setLike(true);
-      postLike(id, userData.token).then((res) => {
-        setLikesArrayLength(res.data.post.likes.length);
-        setActualLikes(res.data.post.likes);
-      });
-      setOnChangePost(true);
-    }
-
-    if (like) {
-      setLike(false);
-      postUnlike(id, userData.token).then((res) => {
-        setLikesArrayLength(res.data.post.likes.length);
-        setActualLikes(res.data.post.likes);
-      });
-      setOnChangePost(false);
-    }
-  }
-
-  useEffect(() => {
-    let toolTipUsersNames;
-    let toolTipUsersIds;
-    let preToolTipMsg;
-
-    toolTipUsersNames = actualLikes.map((names) =>
-      names.username ? names.username : names["user.username"]
-    );
-    toolTipUsersIds = actualLikes.map((ids) => ids.userId);
-
-    if (toolTipUsersNames.length === 1) {
-      preToolTipMsg = toolTipUsersNames[0];
-
-      if (toolTipUsersIds.includes(userData.user.id)) {
-        preToolTipMsg = "Você";
-      }
-    } else if (toolTipUsersNames.length === 2) {
-      preToolTipMsg = `${toolTipUsersNames[0]} e ${toolTipUsersNames[1]}`;
-
-      if (toolTipUsersIds.includes(userData.user.id)) {
-        toolTipUsersIds = toolTipUsersIds.filter(
-          (ids) => ids !== userData.user.id
-        );
-        toolTipUsersNames = likes.filter(
-          (id) => toolTipUsersIds.indexOf(id.userId) > -1
-        );
-        preToolTipMsg = `Você e ${toolTipUsersNames[0]["user.username"]}`;
-      }
-    } else if (toolTipUsersNames.length === 3) {
-      preToolTipMsg = `${toolTipUsersNames[0]}, ${toolTipUsersNames[1]} e outra pessoa`;
-
-      if (toolTipUsersIds.includes(userData.user.id)) {
-        toolTipUsersIds = toolTipUsersIds.filter(
-          (ids) => ids !== userData.user.id
-        );
-        toolTipUsersNames = likes.filter(
-          (id) => toolTipUsersIds.indexOf(id.userId) > -1
-        );
-        preToolTipMsg = `Você, ${toolTipUsersNames[0]["user.username"]} e outra pessoa`;
-      }
-    } else if (toolTipUsersNames.length >= 4) {
-      preToolTipMsg = `${toolTipUsersNames[0]}, ${
-        toolTipUsersNames[1]
-      } e outras ${toolTipUsersNames.length - 2} pessoas`;
-
-      if (toolTipUsersIds.includes(userData.user.id)) {
-        toolTipUsersIds = toolTipUsersIds.filter(
-          (ids) => ids !== userData.user.id
-        );
-        toolTipUsersNames = likes.filter(
-          (id) => toolTipUsersIds.indexOf(id.userId) > -1
-        );
-        preToolTipMsg = `Você, ${
-          toolTipUsersNames[0]["user.username"]
-        } e outras ${toolTipUsersNames.length - 1} pessoas`;
-      }
-    }
-    setToolTipMsg(preToolTipMsg);
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [actualLikes]);
 
   const edit = () => {
     if (editSelected) {
@@ -201,15 +106,7 @@ export default function Post(props) {
           <Link to={`/user/${user.id}`}>
             <img src={avatar} alt="" />
           </Link>
-          <div onClick={isliked} data-tip={toolTipMsg}>
-            {like ? <FaHeart color="red" cursor="pointer" /> : <FiHeart cursor="pointer"/>}
-            <ReactTooltip />
-          </div>
-          <span>
-            {likesArrayLength === 1
-              ? `${likesArrayLength} like`
-              : `${likesArrayLength} likes`}
-          </span>
+          <Likes likes={likes} id={id}/>
           <CommentIcon 
             showComments={showComments} 
             setShowComments={setShowComments}
